requirements.txt
# ========================
# 核心数据层
# ========================
akshare>=1.14.0           # A股/港股/美股数据（港股/美股主数据源 + A股fallback）
baostock>=0.9.0            # A股行情/估值/财务数据（A股主数据源，稳定免费）
pytdx>=1.72               # 通达信实时行情（A股备用数据源）
yfinance>=0.2.38          # 美港股数据备用
pandas>=2.1.0             # 数据处理核心
numpy>=1.26.0             # 数值计算
PyYAML>=6.0.1             # YAML 配置解析
ruamel.yaml>=0.18.0       # 保留注释的 YAML 读写（前端预警管理页用）
requests>=2.31.0          # HTTP 请求

# ========================
# 数据存储
# ========================
SQLAlchemy>=2.0.0         # ORM（MySQL/SQLite）
pymysql>=1.1.0            # MySQL 驱动
openpyxl>=3.1.0           # Excel 读写

# ========================
# 可视化
# ========================
matplotlib>=3.8.0         # 静态图表
plotly>=5.18.0            # 交互式图表
streamlit>=1.29.0         # Web GUI

# ========================
# 技术分析
# ========================
ta>=0.11.0                # 技术指标库（MA/MACD/KDJ/RSI）
scipy>=1.11.0             # 科学计算（分层回测/统计检验）

# ========================
# 策略回测
# ========================
backtrader>=1.9.78.123    # 回测框架

# ========================
# 自动化调度
# ========================
APScheduler>=3.10.4       # 定时任务调度

# ========================
# 报告生成
# ========================
reportlab>=4.0.0          # PDF 生成
python-pptx>=0.6.23       # PPT 自动生成
Pillow>=10.0.0            # 图片处理

# ========================
# 机器学习（AI 扩展预留）
# ========================
scikit-learn>=1.3.0       # 传统 ML（因子选择/模型训练）
# xgboost>=2.0.0          # 梯度提升（可选，按需安装）
# lightgbm>=4.1.0         # 轻量 GBM（可选，按需安装）

# ========================
# 代码规范与测试
# ========================
pytest>=7.4.0             # 单元测试框架
pytest-cov>=4.1.0         # 测试覆盖率
pytest-mock>=3.12.0       # Mock 工具
responses>=0.24.0         # mock HTTP 响应（告警通道测试用）
ruff>=0.1.0               # 代码规范检查（替代 flake8/isort）
